Fix premature settle
The settle criterion did not care about whether or not the sample was at the correct temperature, just
that it was close to the setpoint. Since the LS350 setpoint is constantly changing when ramping we need to add
the criterion that the ramp is done before accepting the sample as settled.